requirements.txt
Django==5.0
Pillow==10.1.0
django-environ==0.11.2
whitenoise==6.6.0
openpyxl==3.1.2
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
